# docs/swagger.yaml
basePath: /
definitions:
...
  dto.RegisterMicroRequest:
    properties:
      chasis:
        type: string
      color:
        type: string
      description:
        type: string
      km_liter:
        type: number
      make_id:
        description: '...o crear uno nuevo (o reusar uno con el mismo nombre+marca)
          bajo el VehicleType "BUS"'
        type: string
      model_id:
        description: Usar un Model ya existente...
        type: string
      model_name:
        type: string
      photo_url:
        type: string
      pin_number:
        type: string
      placa:
        type: string
      route_id:
        description: Asignación a ruta, opcional (se puede hacer después vía /api/vehicle-routes)
        type: string
      user_id:
        type: string
      year:
        type: integer
    required:
    - placa
    - user_id
    type: object
...
  dto.VehicleRouteCreateRequest:
    properties:
      assigned_at:
        type: string
      pin_number:
        type: string
      route_id:
        type: string
      vehicle_id:
        type: string
    required:
    - pin_number
    - route_id
    - vehicle_id
    type: object
  dto.VehicleRouteUpdateRequest:
    properties:
      active:
        type: boolean
      id:
        type: string
      pin_number:
        type: string
      route_id:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RutaSegura API
  version: 0.1.2
paths:
...
  /api/vehicle-routes:
    get:
      consumes:
      - application/json
      description: Get a list of all active vehicle-route assignments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all vehicle-route assignments
      tags:
      - vehicle-routes
    post:
      consumes:
      - application/json
      description: Assign a micro (vehicle) to a route, with the pin number shown
        on the map
      parameters:
      - description: Vehicle Route Create Request
        in: body
        name: vehicle_route
        required: true
        schema:
          $ref: '#/definitions/dto.VehicleRouteCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Assign a vehicle to a route
      tags:
      - vehicle-routes
  /api/vehicle-routes/{id}:
    get:
      consumes:
      - application/json
      description: Get a vehicle-route assignment by ID
      parameters:
      - description: Vehicle Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a vehicle-route assignment by ID
      tags:
      - vehicle-routes
    put:
      consumes:
      - application/json
      description: Update the route, pin number or active flag of an existing assignment
      parameters:
      - description: Vehicle Route ID
        in: path
        name: id
        required: true
        type: string
      - description: Vehicle Route Update Request
        in: body
        name: vehicle_route
        required: true
        schema:
          $ref: '#/definitions/dto.VehicleRouteUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update a vehicle-route assignment
      tags:
      - vehicle-routes
  /api/vehicle-routes/{id}/status:
    patch:
      consumes:
      - application/json
      description: Toggle the status of a vehicle-route assignment
      parameters:
      - description: Vehicle Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Change status of a vehicle-route assignment (soft delete)
      tags:
      - vehicle-routes
  /api/vehicle-routes/register-micro:
    post:
      consumes:
      - application/json
      description: Creates (or reuses) the Model under VehicleType code=BUS, creates
        the Vehicle, and optionally assigns it to a route with a pin number
      parameters:
      - description: Register Micro Request
        in: body
        name: micro
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterMicroRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Register a microbus in one call
      tags:
      - vehicle-routes
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"